+++ docker/docker-entrypoint.sh
@@ -35,19 +35,28 @@ function copyIfMapped {
 
   if [[ "$originalPath" != "$replacedPath" ]]; then
 
-    >&2 echo "copying $fullPath to $TARGET_ROOT$replacedPath"
-
     directory="$(dirname "$TARGET_ROOT$replacedPath")"
     if [[ ! -d "$directory" ]]; then
       >&2 echo "creating intermediate directories for $directory"
       mkdir -p "$directory"
     fi
 
+    newFilename="$TARGET_ROOT$replacedPath"
+
+    # append sha1 sum before file extension
+    if [[ "${APPEND_SHA1_SUM:-false}" == "true" ]]; then
+      sha=$(sha1sum $fullPath | cut -d " " -f 1)
+      # foo.txt -> foo-e1d35a6f7182bfbac1c45f6bfdc12419e06e8d59.txt
+      newFilename=$(echo $newFilename | sed "s/\.[^.]*$/-$sha&/")
+    fi
+
+    >&2 echo "copying $fullPath to $newFilename"
+
     # cp will create all new timestamps for the new file
-    cp "$fullPath" "$TARGET_ROOT$replacedPath.part"
+    cp "$fullPath" "$newFilename.part"
     # copy to a *.part file and rename when complete to make sure
     # watchers will process a complete file (create event is fired before file is complete)
-    mv "$TARGET_ROOT$replacedPath.part" "$TARGET_ROOT$replacedPath"
+    mv "$newFilename.part" "$newFilename"
 
   fi
 }